+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the "timing" extension will be documented in this file. This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+* **DEPRECATED** the `localize` option for advanced `timing.hoverTimestamp.targetFormat`s. Use the new `timezone` option instead.
+* **Added** `timezone` option for advanced `timing.hoverTimestamp.targetFormat`s to enable custom timezone support.
 * **Added** `timing.isoToEpochTimezone` command to enable time conversion including a custom timezone.
 
 ## 2020-03-23 - v2.3.1